netmount client: Configurable response timeout
Adds new arguments `/MIN_RCV_TMO:<seconds>` and `/MAX_RCV_TMO:<seconds>`
to the `mount` command.

The client sends a request and waits for a response for a minimum configured
timeout. If no response is received, it sends the request again and again.
The request can be resent up to 3 times (so it is sent up to 4 times in total).
The timeout doubles with each retry up to the maximum configured timeout.

The minimum and maximum timeout can be configured from 1 to 56 seconds.
The default values are 1 second for the minimum and 5 seconds for the maximum.

Previously, the minimum timeout was hardcoded in code, about 55 - 110
milliseconds, and increased by 55 milliseconds with each retry. Which is usable
on a "fast" network, but generally too aggressive. On a high latency network,
or a very slow network (e.g. RS232 serial line), it can cause unnecessary
repetition of requests and thus more load on the network.