netmount client: Fix: Store packet/request id/seq in resident memory
The resident functions `create_ip` and `send_request` used a static
local variable to remember the last id/sequence number sent. When built,
these variables were stored at the end of the program outside
the resident part. This resulted in a write to memory that no longer
belonged to the `netmount` program. Another program could later be
loaded in this location. And thus `netmount` could overwrite/read
3 bytes of another program.