netmount client: Do not use static local variables in function get_cds
Static local variables were used to cache the state when the function
was first called. Multiple function calls during a single program run
only happened when unmounting multiple disks - `umount /ALL`.

The benefit of caching is not noticeable to the user. It is better
to save a few bytes of memory.