client: Support for sending ARP requests
Previously, the client responded to ARP requests and learned the peer's
HW (MAC) address from them. Until it learned the peer's HW address, it
sent data as a broadcast - destination address FF:FF:FF:FF:FF:FF.

Now if the client does not know the destination HW address, it sends
an ARP request. It sends ARP requests with a 1 second period until it
learns the address (but no more than 5 times).  If it does not find
the address, it uses the original strategy and sends the data
as a broadcast.

A new argument "/NO_ARP_REQUESTS" has been added to disable sending ARP
requests. Then the client behaves as before. It doesn't send ARP
requests, but still responds to them and learns from the queries.

Disabling the sending of ARP requests is useful, for example, for
communication via the class I SLIP packet driver.