client: pktdrv_recv: Better check if receive buffer is free
Previously, the receive buffer was considered used if the length
of the data in it was greater than 0. In fact, the length value can be
negative. The negative value is used as information that the buffer has
already been passed (is used) to the packet driver and we expect
the packet driver to write the received frame into it.

Now the receive buffer is considered used if the length value is
different from 0.

I don't know if there is (and if it is allowed) a packet driver
that would request another buffer to receive data before it fills
the previous one, but this change took care of this case.