server: Using Linux ioctl only in a Linux build (FAT attrs)
Support for storing FAT file attributes is only in the FAT file system.
And it uses Linux ioctl. To be able to build the program for other operating
systems, support for storing FAT attributes is now only in the conditional
build for Linux.

Storing FAT attributes is usually not needed. However, support for storing FAT
attributes for other operating systems, and possibly other filesystems,
may be added later.