server: Use std::filesystem::path::string instead of path::native
In some cases we need to convert `std::filesystem::path` to the
`std::string` type.  The `std::filesystem::path::native` method returns
a native character type. Which is not a problem on POSIX systems where
`path::value_type` is `char`. But for example on MS Windows
`path::value_type` is `wchar_t`. So let's use the
`std::filesystem::path::string` method, which returns the path
converted to `std::string`.

How to convert non-ASCII characters is not addressed.