server: UdpSocket: New method "signal_stop"
In POSIX environment, `select`, `recvfrom` and `sendto` functions are
interrupted when a signal arrives. In MS Windows these functions ignore
signals. In order to be able to interrupt the execution of these
functions, the `signal_stop` method was added. We will call it in case
of signal arrival.

POSIX implementation of `signal_stop` method does nothing. It is not
needed. The functions terminate themselves when a signal arrives.
MS Windows implementation closes the socket and thus terminates
the ongoing `select`, `recvfrom` and `sendto`.