client: Configurable checksums
Adds a new argument `/CHECKSUMS:<names>` to the `mount` command.
`names` is a list of checksums to use. Allowed values ​​are `IP_HEADER`
and `NETMOUNT`. By default, both are used.

The client always sent and validated the IP header and NetMount protocol
checksum. Now it is possible to define a list of checksums to use.
Examples:
`/CHECKSUMS:IP_HEADER` - uses only the checksum of the IP header
`/CHECKSUMS:NETMOUNT`  - uses only the checksum of the NetMount protocol
`/CHECKSUMS:`          - empty list, all checksums are disabled
`/CHECKSUMS:IP_HEADER,NETMOUNT` - uses both checksums - default

Note:
The checksum of the IP header is always sent. It is mandatory. With
the argument, we only disable the validation of the checksum of received
IP headers.