server: Server file names conversion to DOS 8.3
The server implements its own conversion of existing file names to DOS
short names 8.3. The advantage of the conversion is that it is
independent of the operating system and file system. The disadvantage is
that the mapping of converted file names to existing ones is only
temporary in RAM and is unstable (re-created during some operations).

The problem occurs when we have multiple long file names with the same
beginning and then we delete some of them.
Example:
  Initial mapping on the server:
    LONG_N~1.TXT -> long_name1.txt
    LONG_N~2.TXT -> long_name2.txt
  Commands on the client:
    DEL LONG_N~1.TXT
    DIR
  New mapping on the server:
    LONG_N~1.TXT -> long_name2.txt

  So, the file LONG_N~1.TXT (on the server long_name1.txt) has been
  deleted. But after the DIR command (re-creating the mapping),
  LONG_N~2.TXT disappears and LONG_N~1.TXT reappears, now pointing
  to long_name2.txt.

Conversion description:
- Numbers '0' - '9', uppercase ASCII letters 'A' - 'Z' and characters
  '!', '#', '$', '%', '&', ''', '(', ')', '-', '@', '^', '_', '`', '{',
  '}', '~' are preserved
- Lowercase ASCII letters 'a' - 'z' are converted to uppercase ASCII
  letters 'A' - 'Z'.
- Space characters ' ' are preserved except for the trailing ones (in
  DOS they are considered to be padding and not a part of the filename)
- Other characters are omitted

- If any character is omitted based on the above conversion, or the name
  was too long, the name is modified to end with ~<number>.
- If a name collision occurs after lower to upper case conversion,
  the new name is modified to end with ~<number>.
- <number> is a number in the range 1-9999. It starts with 1 for
  a specific prefix and increases by 1.

--------
The server now accepts new optional argument `name_conversion=<method>`
in the shared drive definition. Supported `<method>` are `RAM` and
`OFF`. The default is `RAM`. `OFF` turns off file name conversion.

`OFF` is preferred if we are sure that:
- the files on the server comply with DOS 8.3 rules and
- the file system is case insensitive or the file names are all
  lowercase.

The argument is separated by a comma. If there is a comma in the path
to the shared directory, use the '\' character as an escape.
Do not share the same directory with different `name_conversion` using
the same server process!

Example usage:
netmount-server C=/share/c,name_conversion=OFF D=/data 'G=/share_with\,comma'