server: Implement basic INT2F_UNLOCK_FILE handler
Previously, only the INT2F_LOCK_UNLOCK_FILE function (a client request)
was handled. INT2F_UNLOCK_FILE is now supported as well, but both
handlers currently only validate the file handle and log an ERROR if
the handle is invalid. Actual locking/unlocking is not yet implemented
and may be added later.