server: Add support for setting volume label for shared drives
Previously, shared drives had no volume label. Now, the user can specify
a label for each shared drive. If a volume label is not specified,
the default label "NETMOUNT" is used. To share a drive without a volume
label, an empty string can be passed: `--label=`.

A new option was added:
  <label>=<volume_label>  volume label (first 11 chars used,
                          default: NETMOUNT; use "--label=" to remove)