server: Support Unicode filenames on Windows
Thanks to support for mapping long and non-ASCII filenames to
the DOS 8.3 format, it is possible to share files with non-ASCII names.
The issue occurred when running on Windows, where the `fopen`, `rename`,
`stat`, and `truncate` functions use the local 8-bit encoding for
filenames, which depends on the system locale.

The `rename` and `stat` functions have been replaced with C++
`std::filesystem` equivalents, which work correctly with UTF-16
on Windows. The `fopen` function has been replaced with
the Windows-specific `_wfopen`, which supports UTF-16.
The `truncate` function has been reimplemented specifically for Windows.