client: Make configurable minimum read length for read-ahead
The NetMount client supports read-ahead buffering.

It uses a 64-byte buffer. When reading files, it requests at least 64
bytes from the server to fill this buffer, assuming that subsequent
reads will be sequential and can be served directly from the buffer.
This optimization is only applied when the read request is smaller than
64 bytes.

However, a slowdown may occur if the application performs many small
reads from random offsets that fall outside the buffer range.
To address these special cases, I added a new option to the client
that allows reducing the minimum read length. This setting is
independent for each mounted unit.

A new argument `/MIN_READ_LEN:<length>` was added to the mount command.
`<length>` defines the minimum number of bytes read from the server.
Allowed values are: 0, 1, 2, 4, 8, 16, 32, 64, with the default
being 64.

Read-ahead can be completely disabled by setting the value to 0.