Remove redundant GPL-2.0 license file
At the beginning of the development, I added a LICENSE file with
the GPL-2.0 license text to the project root directory. Later, I added
a new GPL-2.0 license file to the LICENCES directory, planning to store
multiple licenses there in the future. Now I realized that I had already
placed a license file in the root directory at the very start of
the project. I’m now removing this forgotten redundant file.