client: Move most client header files to the include directory
This is a preparation step for an external application that will
implement additional NetMount client commands. The goal is to avoid
increasing the size of the NetMount client and to place extra
functionality that doesn’t need to be part of the client itself into
a separate application that will communicate with the NetMount client.