client: pktdrv_init: Fix releasing packet driver handle for ARP frames
There was a bug in the argument: the type of ARP frame was passed
instead of the handle. The code is executed only in the event of
a packet driver initialization failure, and only when the handler for
ARP frames was successfully registered but the registration of the IPv4
handler subsequently failed. This scenario normally does not occur.