server: time_to_fat: Clamp time_t to valid DOS FAT timestamp range
Added boundary checks when converting time_t to DOS FAT timestamp.
DOS FAT supports times only between 1980-01-01 00:00:00 and
2107-12-31 23:59:58. Previously, times outside this range could result
in incorrect year values while other fields were correct.

Now, the year is checked: if it is before 1980, the timestamp is set
to the minimum valid FAT time (1980-01-01 00:00:00). If it is after
2107, the timestamp is set to the maximum valid FAT time
(2107-12-31 23:59:58).