server: Fix symlink handling for nonexistent target
DOS does not support symlinks. Therefore, the NetMount server follows
symlinks and returns what they point to (a file or a directory).

A problem occurred when a symlink pointed to a nonexistent path. When
building the directory listing, such a link inherited the properties
(size, date and time, attributes) of the previous entry. An attempt
to delete such a symlink resulted in an unhandled exception (server
crash).

Now, a symlink pointing to a nonexistent file is listed as a file with
zero size. Its last modification date is set to 1980-01-01 00:00:00,
the minimum DOS FAT timestamp, and all attributes(ARCHIVE, HIDDEN,
READ ONLY, SYSTEM) are cleared.
Deletion has been fixed, and the symlink is now deleted correctly
whether it points to an existing item or not.